Conflicts:
	src/content_sao.cpp
	src/content_sao.h
	src/script/cpp_api/s_player.cpp
	src/script/cpp_api/s_player.h

+++ src/script/cpp_api/s_player.cpp
@@ -90,8 +90,8 @@ void ScriptApiPlayer::on_rightclickplayer(ServerActiveObject *player,
 	runCallbacks(2, RUN_CALLBACKS_MODE_FIRST);
 }
 
-s16 ScriptApiPlayer::on_player_hpchange(ServerActiveObject *player,
-	s16 hp_change, const PlayerHPChangeReason &reason)
+s32 ScriptApiPlayer::on_player_hpchange(ServerActiveObject *player,
+	s32 hp_change, const PlayerHPChangeReason &reason)
 {
 	SCRIPTAPI_PRECHECKHEADER
 
@@ -208,7 +208,7 @@ void ScriptApiPlayer::on_playerReceiveFields(ServerActiveObject *player,
 {
 	SCRIPTAPI_PRECHECKHEADER
 
-	// Get core.registered_on_chat_messages
+	// Get core.registered_on_player_receive_fields
 	lua_getglobal(L, "core");
 	lua_getfield(L, -1, "registered_on_player_receive_fields");
 	// Call callbacks
